disabled code to unique all conv_op_base_t's seen in rtc
seems to work, but since it doesn't include input and output sizes
maybe it's not too usefull. oops. well, probably a good start for what
we want. maybe we can clean up some of the input-dependant issues
while we're reworking things. perhaps we should move the putting of
the in/out dims into the dims_val from the rtc code into the conv_util
level ...