rtc: fix k1conv write_xposed optimization finally/hopefully
OMG that was broken a long time. since the bug was that it was
effectiely disabled, and it only effects a single test case, and
causes only a small perf regression, i guess it was easy to miss. but,
it's a placeholder for more complex optimizations so we want to make a
point of keeping it working. we need a testcase that checks that the
optimization is used somehow ...