build: tweak path handling for obj_list in boda_make_util.py
rather than looking in '../obj/filename', it now just looks in 'filename', so that the build process is agnostic toward
the name of the 'obj' directory. in particular, if you make a new obj/lib set of directories, and symlink the makefile
and obj_list into the new obj directory (e.g. obj2), and then use TARGET=../lib2/boda when running make, you can have 2
builds of boda living in the same working copy.