build: clean up version of obj_list (obj/obj_list.android)
by (correctly?) using -rpath-link, we don't need to manually list the
sub-deps of libOpenCL.so; this removes a ton of crud from the link and
now the only difference between the msm8994 and msm8996 builds is the
SYSROOT path. so, we remove the msm8994 varaint of the obj_list for
now in favor of just having a comment-out version of the right
SYSROOT. note that SYSROOT can be passed as an option to make as one
way to override the default.