add intermediate class conv_op_base_t in between op_base_t and conv_op_t
this class is op_base_t + a coi field. also, we remove op_info_t
(leaving it as a typedef for now to avoid some churn), and make
op_info_t::init() be a free function
add_cnn_codegen_annotations(). this new function takes and operates on
a conv_op_base_t, as it does not need/want the arg_map or knowledge of
what global variables the operation will act on. so, for the cnn_prof
mode, we can now load conv_op_base_t's instead of conv_op_t's with
empty/bogus tops/bots/arg_maps.