remove rtc_func_sig_t and replace with op_base_t
rtc_func_sig_t previously derived from op_base_t and added almost
nothing: one duplicated-with-base member func, and one other func that
was made into a free func. we avoided this change earlier to
maybe-avoid churn, but it's in the way a little now, and we havn't
touched that section in a little while anyway.