sgemm: prep for actually starting basic boda version
in particular, we changed the arg names from a,bt --> a,b *but* we
still allow a or b to be transposed on input (as specified by the CUCL
dims, although this isn't checked/handled by the culibs wrapper
yet). and, currently, we have 'a' transposed, but *not* b, which is
the reverse of what we had before, but seems like the easiest starting
point. note that (as expected?) cublas seems happy with this format,
perhaps happier that the prior one.