first pass at boda sgemm
block size 64 only with 64-threads, 64 outputs per thread. gets 40%
eff of titan X, not too bad. note that block size 128 kernel should be
better (less/not BW limited), but is in fact considerably worse. this
is similar to prior experiences where large threads-per-block is
'bad'. anyway, there's the usual list of debugging and such, but this
seems pretty good for a first pass: ~3 hours work, seems correct
compared against cuBLAS on various square inputs with random data, and
gets within about 2X perf (and would be expected to be BW limited at
not-too-much-more-than-that anyway).