simplify config file a bit and make optional; remove unneeded canonical() path calls
in particular, deployment to android just for running
ipc_compute_worker will run without a boda_cfg.xml file or stub test
directory now. so, the android binary is mostly standalone for some
uses cases now. but libgnustl_shared.so is still needed +
LD_LIBRARY_PATH to load it.

note that the config file being optional will cause a different
(worse/less-intuitive) failure mode when trying to run the tests
without a valid config file, and there's no feedback when boda is
reading (or not reading) the config file. so that's not good.