sgemm: add no-local-mem variant, and perf #s for running it on SD820
note that it seems that c_r[] may not be in registers for t_tile_sz=8
/ sgemm_bsz=16, but it's unclear. SDP static-kernel-analysis lists
'Scratch space : 256' for these kernels, which seems to be the 64
elements of c_r[], but it's unclear where they actually live. and of
course, it's not clear if the static analysis is reliably for the
actual kernel/compilation used at runtime. the Adreno profiler had
static analysis per run kernel, but SDP doesn't seem to, so you have
to manually compile (including setting options) each kernel for
analysis, hoping you get the right source code and options ...