in preparation for more SD820 local-memory experiments, more work nVidia/Titan-X sgemm
add inner-loop unrolling, which (as expected from experience with
convolutions) yields significant speedup. note that we still lag
cuBLAS, but not as badly in the (very limited set of) square
power-of-two cases we're testing on.

next TODO is to refactor the sgemm tuning parameter passing in
cnn-prof since we have so many of them now.