sgemm: add/fix support for inner loop unrolling in sgemm_no_local.
note: doesn't actually seem useful on nVidia or SD820: nVidia is
marginially faster with N>1, but SD820 gets much slower. hmm. i guess
i didn't have any opinion on how this would work in this case, and i'm
not sure it much matters. but it doesn't exactly bode well for
unrolling on the SD820 in general.