sgemm: add some automated/unit tests for sgemm in a few modes for 512x512 only
better than nothing. note that we are hampered by the high (~300ms)
nvrtc per-compilation-unit time overhead for our data generation and
stub functions when using culibs. otherwise, in thoery, these tests
should be pretty fast, since the OpenCL compile overhead isn't too
high (a few ms). we might adress this in a few ways, but we'll leave
it alone for now.