conv: inching toward cudnn compare support
still can't quite get to cudnn conv() call in culibs-wrap yet; need to
deal with issue of incompatible filter layouts. probably best to
generate 'normal' filters, and then insert the needed filts xpose
somehow? we've considered folding this issue into codegen before, but
maybe we can hack around it for yet longer somehow? in general, this
is related tot the 'gap' code in rtc-fwd.cc that is post-annotation
but pre-call, and sometimes inserts multiple calls for one operations.