cnn-prof: add hacky support for filter xpose
as opposed to the prior method of directly generating xposed
filters. now, the gen_data kernel always generates data in 'reference'
format, and the normal xpose_filts call is inserted if required. this
allows the use of one gen_data function for all conv operations, even
if they need xposed filters. finally, we can get to the TODO assertion
in the cudnn wrapper code.