rtc: put rtc ref inside codegen
codegen is now one step closer to being a full wrapper/layer over the
rtc_compute_t layer. we could take it a bit further (and maybe we
should), but for now we'll pause and attempt the desired change of
passing the annoted-and-reduced op_base_t from codegen::compile() into
rtc::compile() (or something similar) so we can get the op down into
where we need it for the culibs wrapper.