conv: k1conv_simd debugging state
ifdefs control fast/wrong vs. slow/correct output writing. still
unclear what issue is, but based on static analysis there seems to be
unexpectly large differences in the # os instructions and # of waves
between the two versions. of course it's possible the 'fast'/hacked
version is allowing some invalid optimizations to the compute section,
but on the other hand the 'fast' perf is similar to the sgemm_simd
perf for the same input size.