k1conv_simd: use chan:pel format instead of chan:img:y:x for in/out
it's all that much closer to matching the sgemm now, but more
importantly, we can pad the input more easily and effectively. instead
of adding an extra whole image, we just add any needed pels. this also
means the size of pels (and thus the chan stride) is an exactly
multiple of all the work sizes, which can be important/needed/helpful.