cnn-prof/codegen: add minimal/experimental support for per-variant tuning specification
since we now have no single set of tuning params that allows running
conv-ops-small, we need at least something in this direction. in
particular, due to the SD820 hardware limits / codegen-limitations, we
routinely will hit the max-workgroup-size limit at runtime for various
configurations. really, it seems like the hardware should allow these
cases to run (albeit slowly), but i guess the spec is unclear to me on
that. anyway, such cases are to be avoided in any event, but it's a
hastle to deal-with/avoid the run-time failure cases.