conv: moe work on conv_simd. currently might be ~s1conv_simd.
broken intermediate commit. conv_simd may actually be working/complete
for cases with stride=1 -- it partially handles strides, but not
enough to be right. however a couple cases with stride=1 but ksz>1 do
seem to work. so, we might actually want to return to this version
later if we need/want to special case stride=1.