conv: working toward bigger sweep of k1conv_simd/conv_simd on SD820.
getting hung up on random build failures with CL_OUT_OF_HOST_MEMORY,
which seems to indicate some register or sm limit being hit in the
compile. sigh. smaller register blocking 'fixes' these cases, but
right now it's hard to figure out which cases need it or how to detect
that. it seems like we need to do some work on returning errors over
IPC and catching them in cnn-prof.