cnn-prof: change eff output in sgemm mode to be comparison-table-format
this drops the %peak column, and adds the comparison runtime/fps and a
speedup column. this is the format we're currently using for the
tables in the paper. we can generate the cuBLAS/boda comparison
directly, but since we don't have a android-ocl-QBLAS wrapper (similar
to culibs-wrap), we can't auto-general the QBLAS/boda comparison for
now. since it's a simple case, though, we can just run QBLAS-test
manually a few times to get the timing and fill it in manually. at
some point an official wrapper might be nice/useful, though.