broken commit: rtc_compute iface updated to use nda_t instead of nda_float_t
this (naturally) breaks rtc_fwd (and rtc_prof), and maybe other stuff
that somehow doesn't show up. we'll tackle those changes as another
commit. note, however, that in many places it should (in thoery) be
okay to pass nda_float_t where nda_t is expected, both for read and
write, since and nda_float_t is just an nda_t that happens to have
tn="float" (and related C++ level iface functions), but is a perfectly
fine nda_t.