gather all template var names during CUCL template parsing
also, refactor per-line processing into seperate function, with better
error reporting via catch-and-annotate in the caller. note that now
could(/should?) record the post-parsed template string to perhaps make
the later actual substituion more effecient(/cleaner?), but we don't
try to do that yet (we still call to the 'regular' template
substitution function). eventually, we'd like to ensure that the later
substition cannot fail by pre-checking the existance of all needed
template vars, but we're not there yet. but capturing all the names is
the first step.