add reshape/view rtc func and use it to fix breakage of prior commit
now we can properly make the call-time arg checking fatal and always
enabled, and this fixes the cucl_arg_info in the var_stats case. this
allows us to finally remove the u32_arg for the input size in
var_stats, in favor of using a CUCL IX decl on one of the (now
dynamic) input vars. we also need to allow setting of %(tbp) via a
magic element of str_vals ("tpb"), as well as to allow the magic
GLOB_ID_1D CUCL IX special case to use a pre-set valud for tpb. yep
all that to remove *one* of the u32 args for var_stats! sigh.