experimental changes/tweaks to support make_scalar_nda<T>()
this function is designed to easily return a p_nda_t of size 1 of any
type. the juggling/tweaks to the ndat_info_t handling are hopefully
not too important, but seemed to be an inprovement. we didn't/don't
have much type-generic nda_t functionality in general, so it's no
surprise we'll need to churn the existing stuff as we add more. this
change is designed to enable using nda_t instead of a fixed uint32_t
type for the pass-by-value CUCL args.