attempt to clean up MULTI handling
perhaps partially successful, but considerable stymied by the
inability to pass pointers-to-pointers or pointers-inside-structs in
OpenCL. it's unclear if CUDA has the same issue, or if we'd maybe want
to hack around this somehow. but for now, we leave most of the code in
place that would allow us to employ structs-with-multiple-pointers,
and arguably the other multi hanlding changes are valid cleanup. note
that, currently, *only reduce* actually uses MULTI support. it seems
like a lot of complexity for that one case, but it's not clear what
else could be done, or if maybe other similar cases will occur later.

on a related note, if we move to our own allocation handling, maybe
using a single arena, we could use struct-of-indexes for the multi
case.