minor tweaks to android compile/rtc_test example lines in notes
since we seem to be putting boda in /data/local/bin most recently,
instead of in its own tree. however, i'm thinking we'd actually want
to go back to having a boda WC-like tree if we're planning on running
more of the framework on the device anyway, since it doesn't seem nice
to, say, put testing data in /data/local/test ... hmm.