oops, add special cases to do-nothing for no-functions-to-compile case
we're doing some belt-and-suspenders here with having checks at both
the rtc_compute_t level and the rtc_codegen_t level, but that seems
right/okay to me in this case. note that, without this fix, we were
doing an 'empty' nvrtc compile (no funcs) at *every call*. that was no
good! the tests never passed. but since this is interesting/seperate
as a change, we'll resist the urge to rebase this incident into
oblivion.