rtc_compute_t: clean-up arg handling: remove nda_args, cucl_arg_info; make args vector of rtc_arg_t
where rtc_arg_t specifies each are as either a name of some rtc var
(to pass as a pointer) or a nda value (to pass flat by value). also,
remove the extra level of vector-ness that was intented to support
multis. for now, we're just accepting that the argument list at the
rtc level is flat, ordered, and not well checked (and in turn defering
yet more to the rtc_codegen_t level, and assuming there will be few
direct users of the rtc_compute_t level). all args passed are now
declared CUCL args, and can be in any order (meaning an arbitrary
order/mixing of by-value and non-by-value args, but this is untested)
but ARGINFO is still specially handled still and must be last.