remove CUCL ARGINFO directive and use 'regular' CUCL IN for cucl_arg_info arg instead
however, we now need various special-cases on the *name* of the
agument being exactly 'cucl_arg_info'. in particular, in about ~4
places we need to skip it, and in a couple others we need to put the
actual cucl_arg_info handling. the rtc_codegen_t layer magically adds
the argument at call-time (like it did before), but now it's a bit
less special cased (at least in some ways). in particular, it can now
go anywhere in the argument list. we don't/didn't care too much about
that, but it makes our goal of uniform arugment handling that much
closer. and, before, we were unable to even error-check if the
cucl_arg_info arg wasn't last, so that wasn't great.