nesi: first pass at nda_t support complete
whew! yeah, i can see why i been putting that off. hopefully it will
both work and be useful. note that we now finally/sort-of have
(limited0 support for vectors in NESI, in the form of ndas. of course
this is far from general vector support, but for the cases where we
want to specifcy something like a nda, it should be okay.