reduce external usages of op_base_t::dims_vals via getters/setters
who doesn't love getters and setters, right? a big of a monkey-change,
and some significant churn, but not too bad. the idea here is to ease
experimentation with nda_vals, but it's a bit experimental. the change
doesn't seem too bad (other than having to do it), so that's nice.