broken commit; working towards actually replacing dims_vals with nda_vals
although not trying to actually use it as anything other than storage
for dims yet. there were a number of derived-and-base class usages of
dims_vals that our last set of changes didn't pick up, and those are
changed here. it would have been better to catch those earlier;
oops. also, we needed to add more support for null ndas, and there was
a wrinkle with trying to use an unitialized dims_t() as a special
value for the 'null arg' case, which no longer works if the dims_t
must always be dims_t::valid(). we also needed to add a none_ndat for
the "none" type, althogh we don't (and can't) support none/void in
nda_dispatch (which in turn needed some refactoring, but it okay since
we don't really need to nda_dispatch() unless there are elements).

things aren't totally broken, but we currently get the 'ol
no-fun-to-debug vector-range-check top-level exception on full-stack
tests, and it's not too clear where to start looking ...