be a bit more careful casting template string constants to float
it seems a few intermediate calculations might have been using double,
which slightly changes the results. using float, the errors on a
couple layers (norm2/pool2) are a tad higher for googlenet. of course,
there may still be underlying/other issues with LRN as noted in prior
commits when we were dealing with it. but, for now, we accept the tiny
bump in tolerance and move on for now, remembering that at some point
we should maybe revisit/resolve the LRN issue.