more op_base_t refactoring: remove type, replace with "type" str_vals field
a bit questionable ... but let's give it a shot for now. we're
planning to move most of the existing str_vals into nda_vals, but
while we're at it, we're trying to be clear is we really need str_vals
at all, and if so, for what. currently, it looks like "type" and
"func_name" are two key, important str_vals. now, we could have
explicity string fields for both (like type was), or niether (like it
is now), but it's not clear really why we'd want one to be a top-level
field and the other to be a str_val? hmm.