port most usages of str_vals to nda_vals
that got out of control. there are a few test updates in here as well
since the op_base_t encoding/formats changed again. we also some
oddities like the sgemm functions taking a 'vw' uint32_t argument
... that of course must be constant, and is treated as a string, since
it's used to declare types and such. in general, we havn't dealt with
the issue of gen-time vs. call-time scalars, but maybe that will be
more natually part of making DYN refs work anyway (which, in thoery,
this entire branch is a preperatory refactoring for).