move call_id out of various structs and make it a return value
we didn't want to churn the ifaces too much by doing this earlier, but
it seems like a nice minor cleanup now, and the ifaces in question
seem pretty stable. there doesn't seem to be any reason to think we'll
need more/differnt return info from the 'run()' call, but i guess at
the least maybe splitting the input and output parts of run() is a
good idea. since the only output is the uint32_t call_id right now, we
just return that without making some new struct.