refactor rtc_codegen_t gen_func() interface
two changes here: (1) firstly, we return an p_rcg_func_call_t instead
of a p_rcg_call_gen_t from gen_func() now. the idea is that we may
want to pre-fill-in some parts of the arg_map for information that was
provided as part of the full operation but got reduced out. in
paticular, the goal is that this will enable DYN REF args. it should
also enable "dynamic scalars", which we don't have a good term for
yet, and can't be specified at the CUCL level yet. (2) we pass an
intial arg_map into gen_func() -- most callers call gen_func() and
then immediately create an rfc from the result with an on-hand
arg-map. that is, they have all needed info to make a call at gen
time. so, maybe we should allow/require the caller to provide all info
needed for the call at gen time, on the thoery maybe we don't really
need a gen/call split. that's part of this commit is more experimental
and maybe less important.

note: two tests broken; seems related to apply-xpose (which was indeed
non-trivally tweaked).