fix gen_apply_func_to_var() bugs by reverting to generating the function first
and then using the gen'd function name to unique the ret_var. in
particular, this will indirectly unique ret_dims (and any parts of oi
that affect the result?). anyway, we did have a FIXME there, and the
call-time dims checking caught the error cleanly enough. this case,
like in rtc_prof, is one of a couple where we can't easily get the
arg_map before generating the function, because the arg_map depends on
the details of the generated function. as mentioned, passing in an
'early' arg_map is optional, and maybe not an important thing to
do/allow ...