remove "tpb" from always_keep hack
it's easy enough to add an explicit tpb uint32_t argument to var_stats
CUCL template; as will all the scalar args, we still need to deal with
gen-time-const scalars (vs not-gen-time-const), but otherwise it's
fine. should be okay for now.

the bulk of the change is moving a block of code from
instantiate_template() to init() so it can happen before setting "tpb"
into tsvs, which now much be guarded, since var_stats explicitly
declares "tpb", and such the arg-processing code has already inserted
it into tsvs.