pass DYN REF dims to rtc_compute_t layer from rtc_codegen_t layer (via arg_map)
the code's starting to get a bit messy; we must be getting close to
success now, right? on a related note, maybe we can start treating REF
args as passing null pointers soon? if that actually seems like a good
idea, i dunno ... still need to deal with call-constant-vs-not-scalars
issue too ...