fixes to caffe_fwd for new version of caffe
currently, our update net code is out-of-date, and doesn't contain the
'legacy' input to input layer conversion code. at some point, we
should probably port that, and change our front-end to handle the new
input layer. however, since we normally read train-val style nets with
data layers (not input layers), it may not matter too much, at least
right now. certainly, we'd need new-style nets we wanted to read
before it much mattered. but in the mean time, we shouldn't be
generating the legacy input params in caffe_fwd anyway; it's not like
we're writing that net out, it only+must work with the current
caffe. so we adjust it to emit the new-style input layers.

after that, the tests seem to run to completion anyway, but there are
lots of diffs. maybe it's all just tolerance issues, or maybe things
are broken. since both caffe and cudnn have changed/upgraded, there's
no lack of possible sources of issues.