make REF 'real' args (passed as null pointers at call-time); use no-data-ndas for non-REF NULL/optional args
notes: (1) the only optional args currently are the ones generated by
some ops that are only sometimes needed: out_in_xy for pooling, and
out_scale_base for LRN. these are only needed/wanted when bck ops are
added, and then they represent an optimization. (2) the bulk of the
change lines are updating all the CUCL templates to the new REF
syntax. note that some (untested) CUCL templates are now stale, and
will be updated in a sepeate commit to isolate the changes to untested
templates. note also that there are still a few bits of "REF" specific
code, but REF handling does seem better (less special) overall.