use minimal subset of pascal VOC2007 for testing
we accomplish this by making the default location for the VOC2007 data
be bench/VOCdevkit, and then make that directory be a
gitlink/submodule to a fork/copy of the needed subset of the pascal
VOC2007 data hosted in a repo that is expected to be a sibling of the
boda repo. for now, i'll provide a suitable repo alongside the boda
repo on github, but bear in mind that, as with all submodules, the
linked-to repo should not be assumed to be part of boda, and may not
exist (i.e. user must provide thier own), or may have a different
license.