make default models location be inside boda directions (aka WC/tree)
also, add utility to upload/download models from b2. this is just a
first pass, and downloads are limited to 1G/day for free. like with
the pascal dataset, we should really find a way to reduce the size of
the needed models for testing. the current set is ~850MB. much of that
is 3 large alexnet-like models, and are probably mostly duplicated
with each other. further, for testing, it may not much matter what the
models are, so some highly-lossily-compressed versions, maybe with
some deterministic randomness added back in at decompression time,
might work just fine. however, it's not exactly on the critical path
at the moment to deal with this. there's probably some low hanging
fruit, but it's not clear exactly what it is ...