culibs-wrap: make cu_work scratch-space-var be persistent and remove dup'd CUDA alloc code
this was harder before: (1) we decided to put an rtc reference in the
culibs wrapper object (2) we had easy allocation of uint8_t-vector
vars and, (3) had a way to get raw CUDA device pointers out of the
rtc. but now that we have all three things, it's easy enough. note:
not tested by any automated tests. we ran a single cudnn-enabled
test_compute mode call for googlenet-v1 to test.