test_nesi: move needed-features test prior to cli_str processing
as expected, our needed-features processing doesn't help when the NESI
init for the entire test fails. so, for non-cli-str-based tests it's
not so useful. this change is a compromise: the behavior of the
missing-feature test now differs between the cli-str and non-cli-str
cases, but on the bright side it does what we want in the cli-str case
(where the init of the mode/cli is/can-be deffered until after the
cmd_test_t NESI object itself is init. so, in short, we need to update
any tests that want to use the 'needed' feature to use cli_str
format. since there seems to be little reason not to do that (other
than the hastle), this seems okay. maybe eventually non-cli-str format
test support will be removed.