mark CUDA-specific rtc test as needing nvrtc
this test explicity (1) selects the nvrtc rtc backend and (2) passes
CUDA-specific code (not portable CUCL) to the rtc_compute layer. so,
there's no point in running it if nvrtc support isn't availible. for
many other tests that use the nvrtc backend, they will still work even
if it's aliased to OpenCL, since they use portable CUCL (i.e. don't
really care about the backend for functional correctness). this test
is (intentionally) an expception. similarly, tests that use culibs are
probably in the same boat and should be so marked, but we've perhaps
got bigger issue there, so that's TODO.