experimenting with various methods/approaches to generic-ification of code using p_nda_float_t
much of this code can use p_nda_t instead, but internally downcast to
p_nda_float_t -- this fixes the interface, and doesn't junk-up/change
the code much, but will fail at runtime for non-float NDAs. in other
case, we can use nda_dispatch() and make the operation actually
generic. the downside is that we don't have a syntactially nice way of
doing this: it minally requires a functor and a wrapper function, and
lots of duplications of arg names/passing. so maybe we don't want to
do that too broadly yet if we can avoid it (and maybe do something
better later).

for comp_vars(), we already had a generic version, but we needed (1)
some helper glue to deal with the containter-of-dervied issue, and (2)
some common-denominator way of printing/dealing with deltas across
types. this commit fixes (1), but not (2), so test_quantize is broken
for the moment.