stub for digest-output from test_compute_t
a start on enabling use of digests for comparison rather than running
a comparison engine. the idea is that most tests will be ported to use
this method, which allows each case to be tested against a known-good
result rather than a hopefully-correct reference engine. in
particular, this means that the tests should be meaningfully runnable
even when only a single method for running a certain test is
availible. also, having at least some tests with golden results should
catch some forms of errors that would otherwise be uncaught (e.g. if a
test decided to start running on all zeros as input).