test_compute: add cmp_digests option
and add underlying mrd_comp() function for digests. with cmp_digests
enabled, all tests pass except for the expected two diffs for the
quantize test (because the digests differ as expected, like the
non-digest comp does now), and for the nin test (since this had
warnings about skipping comp for different-sized vars, and now has
extra similar digest comp warnings/errors for the same vars.

note the large comment related to adj_mrd for cases with large #s of
subsamples. this was expected, but we're not too happy about it. note
that, without the adjustment, only 4 tests fail: they are all grad
tests, and the failures are all for stride=1/2 checksums on
largeish-vars. so, despite making-sense and being somewhat nerfed, the
adj_mrd change still feels a little too anti-conservative.

next TODO seems to be to add support for reading/writing known-correct
digests to test_compute_t and/or comp_vars, and then to actually use
that support to port all the existing tests to using digests instead
of live comparisons.