nda_digest_t: more work on test_compute_t write_digests=1 mode
seems to work now, but still unused/untested in automated
tests. currently, the .boda stream diff for nda_digest_t's uses an
mrd of 0.0, which means the digests must be exactly equal. in
many/most cases, if we generate a digest, it should indeed be
identical if we re-generate it. however, for some cases, this is not
true (i.e. non-deterministic runs). in those cases, we need to use
some error tolerance.

bearing this in mind, we should be able to now add regression-style
tests for all currently-run cases. then, we need to go back and add
cases that take all the pairs of digests that should be 'almost equal'
and test that. or something. it's still a little unclear how to get
from here to there or exactly where there is ... but we're getting
closer i think.