nesi: filename_t dump: add commented-out version to dump expanded filename instead of in
in general, there's some maybe-expectation that, if you print a NESI
struct, you can get back the same struct by reading in the resultant
string. this is certainly not actually true in general, and the
filename_t case is a particular issue. in general, we expect that we
should print the value of a field if it is 'non-default', but for
filename_t, using the 'in' value isn't really right, since the
expanded version can including data from external strings (at any
level of the input lexp). so, maybe we should print the 'exp' version
-- it's legal to put that back in as 'in', more or less, although we
should escape any '%'s. ehh, it's not to critical at the moment.