working toward digest-based test_compute
note that we're also planning to move to generating the set of
test_compute tests on the fly, which should ease dealing with
different backends being enabled/disabled. also, we're planning to
modify test_compute to test all backends for a given input/model
combination at once, rather than only in pairs.