add test_compute_multi tests.
these test should subsume all the 'test_compute_*' and 'test_grad_*'
tests in test_cmds.xml (which all use the test_compute_t mode). but,
in this commit, all tests are still present. the downside is all the
dup'd code and the extra (almost double) test run time. we'll remove
the dup'd tests next, and try to verify carefully that the new tests
really are working and cover the old ones next.

after that, there are a few more tests that use test_compute_t that
need to be migrated or otherwise dealt with, and then the old
test_compute_t mode can be removed.

finally, we need to add some vs-known-good-digest mode/option to
test_compute_multi_t, in the case where we want to compare one
backend's results against a different backend without actually running
the comparison backend. in the limit, this allows adding a new test
case where only one backend can be run at a time (i.e. when we can
never run multiple backends at the same time for comparison testing,
or at least when we can run some backend only by itself).