add single-image-image-as-input support to capture_classify
this is almost what cnet_predict_t does already, but in this case we
get the GUI display window with the annotations overlayed on the image
(instead of printed to the terminal). not the cleanest, but not too
horrible. since the capture isn't started/setup in this case, it
should work (as one would expect) when no camera is present, so that's
nice for testing too. again, not as modular/clean as it could be, but
okay for now i guess.