working on allowing test_compute_multi tests to pass for subsets
that is, when running a subset of the full/commited-known-working
backends, we'd like the test to pass. for this, we need all the
digests to be correct (or have some way to ignore missing digests if
they were not run). also, we need the main text output to not depend
on the set of backends runs. this commit removes the redundant
per-backend printout of the list of vars to compare.

of course, this makes any errors harder to read, but we're free to
improve the error-case output text without impacting the
known-good-case output text, which should more-or-less just always be
'all is well'.