test_compute_multi: compare digests against passed-as-input known-good digest stream
instead of comparing the digests of backend[i] with backend[0]. this
generally does about the same thing, since when we generate the tests,
we pass in the caffe digest (which is/was backend[0]) as the
known-good digest stream. however, note that we're using the on-disk
copy of the digest, not the newly-created one. so, this works even if
we don't run the backend[0]/caffe, and doesn't depend on which backend
is [0]. note that we still compare the full-ndas of all backends[i]
against backend[0], though. however, we now have both regression and
comparison testing for the case of running a single backend, which is
what we wanted.