test_compute_multi: fix typo in diff str; comp be[0] vs be[0] for consistency
the typo was that we printed cfn[0] instead of the name of the
known-good digest file when diffing against the digest (just a failure
to update the message wrt the prior commit). normally this wouldn't be
in the tests, but for now it shows up for the cant-compare-due-to-size
cases in nin (which is ... maybe good since it caught this?).

still todo is making the tests pass with some backends disabled; these
fixes are things we turned up when trying to do that. the comp be[0]
vs be[0] thing means that all the per-backend outputs have the
vars-to-compare line now. bear in mind, though, that we'll still
compare against be[0], which will depend on which be[]s are run, and
we don't print anything out about that, which is a bit
sketchy/confusing. but if there are no diffs, the files will be
consistent anyway ...