test_nesi/test_compute_multi: add ignore_missing_outputs per-cmd-test option; use it for gen'd test_compute_multi tests
now, if we run with caffe disabled, we should only get a
warning. semi-tested by hard-disabling caffe in the test gen code, but
not yet tested by actually disabling caffe in the build. it's maybe
too soon to do that, since while test_compute_multi may be okay, we
need to fix some other tests first, and we probably need some
'default'/'use-any' backend mode for various tests for disabling caffe
to really be sensible yet.