test_compute_multi: make kg_digests_fn be optional
but, warn for each backend/var that a known-good digest comparison was
not performed. for some tests, it might actually make sense to just
not do a digest comparison, since for a one-backend regression test,
it's somewhat silly: the comparison is just dup'd with that performed
by the test_cmd testing harness. but if we do force the known-good
digest to be specified, we can run the test standalone, as well as
changing the backend, and we still get a comparison. so, for now we're
sticking to our (somewhat redundant) guns.