test_cmds: convert test_quantize_1 to use test_compute_multi
it's now a regression test, and we lost the nice output that showed
correct quantization. but, it's okay for now under the assumption the
output didn't change during the conversion (not the best assumption,
granted). we can maybe add-back some optional/comment-out version that
produced the quantized/non-quantized comparison (that is expected to
fail).