working toward dynamic-default for rtc_fwd backend
instead of hard-coding nvrtc, we leave the default rtc
null/unspecified, and manually create it in code if it's null. for now
we're just using the same string to create it that was the default, so
no behavior should change. but now it should be easy to add logic to
pick an enabled backed. for now it's pretty special-cased, but maybe
that's okay.