make default conv_fwd for run_cnet_t be rtc instead of caffe
maybe questionable, but for portability, this is better for
now. perahps we really need the default to be null/unspecified, and
add logic to set a reasonable default when it's null when needed (as
we did with the rtc_fwd backend selection). however, it seems a little
scary to try to update all the usages at the moment? we'll see how it
goes.

on a related note, it seems only the cnn_op_info_t tests need
attention in the no-nvrtc/no-caffe caffe case. of course we don't want
to just mark them needs_nvrtc, but maybe they don't need caffe, so
they're not too hard to fix? on the other hand, it's quite possible
that the real fix it that the entire cnn_op_info flow needs the same
treatement as we gave to test_compute_multi. sigh. maybe we can put it
off for a bit.