new mode: op-info; mostly copy of cnn_op_info to start
the plan is to split apart the latex/report-generation and op-running
parts of cnn_op_info. the new op-info mode will be just the op-running
part. also, we're making it take a variable number of backends instead
of exactly two, similarly to what we did with test_compute_multi
vs. test_compute.

for now, op-info will still do error-checking between all engines, but
as with test_compute_multi, we may need/want a vs-digest comparison
mode. we'll need to add digest read/write first for that to work of
course.

currently op-info compiles but is untested (and presumably
broken/incomplete). but the test still pass, so hopefully the various
refactorings that are part of this commit didn't break anything:
moving the sgemm annotation code to cnn_op.cc, and adding a few
typedefs here and there.