ops-prof: allow some types of annotation failures and add error to op_tune_wisdom_t
in particular, trying to use conv_simd with Kb!=1 fails an assert(),
but we converted it to an unsup_err() which is caught, and just that
parituclar case is marked as failing. as with most such cases, it's
really (much!) better to 'fix' the issue, but we probably need a
general way to handle such unsupported cases, so this seemed like as
good a time as any to add such support to ops-prof. it turned out that
that was the only failure in the 1-5-20-alex-nin-gn conv ops set with
our current set of op-tunes. really, until we go back and run things
on the SD820 again using the new ops-prof mode, we probably won't
shake out all the issues.